lyrics: split model request and cache modules
