+++ crates/utils/src/lyrics.rs
@@ -1,9 +1,19 @@
 use percent_encoding::NON_ALPHANUMERIC;
-use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet, VecDeque, hash_map::DefaultHasher};
+use serde::Deserialize;
+use std::collections::{HashSet, hash_map::DefaultHasher};
 use std::hash::{Hash, Hasher};
 use std::sync::{Mutex, OnceLock};
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+
+mod cache;
+mod model;
+mod request;
+
+use cache::{
+    LyricsInflightGuard, load_persisted_lyrics, lyrics_cache, store_lyrics, try_begin_lyrics_fetch,
+};
+pub use model::{LyricChunk, LyricLine, Lyrics};
+pub use request::{LyricsRequest, LyricsServerAuth};
 
 #[derive(Debug, Deserialize)]
 struct LrcLibResponse {
@@ -13,112 +23,6 @@ struct LrcLibResponse {
     plain_lyrics: Option<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct LyricChunk {
-    /// A timed lyric chunk. Most providers use whole words; Apple Music can
-    /// return smaller syllable-level chunks.
-    pub start_time: f64,
-    pub text: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct LyricLine {
-    pub start_time: f64,
-    #[serde(default)]
-    pub end_time: Option<f64>,
-    pub text: String,
-    #[serde(default)]
-    pub chunks: Vec<LyricChunk>,
-    #[serde(default)]
-    pub parent_line_index: Option<usize>,
-    #[serde(default)]
-    pub background: bool,
-    #[serde(default)]
-    pub opposite_turn: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Lyrics {
-    Synced(Vec<LyricLine>),
-    Plain(String),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LyricsServerAuth {
-    pub url: String,
-    pub token: Option<String>,
-    pub user_id: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LyricsRequest {
-    pub artist: String,
-    pub title: String,
-    pub album: String,
-    pub duration: u64,
-    pub track_path: String,
-    pub server: Option<LyricsServerAuth>,
-    pub prefer_local: bool,
-    pub enable_musixmatch: bool,
-}
-
-impl LyricsRequest {
-    pub fn new(
-        artist: impl Into<String>,
-        title: impl Into<String>,
-        album: impl Into<String>,
-        duration: u64,
-        track_path: impl Into<String>,
-    ) -> Self {
-        Self {
-            artist: artist.into(),
-            title: title.into(),
-            album: album.into(),
-            duration,
-            track_path: track_path.into(),
-            server: None,
-            prefer_local: false,
-            enable_musixmatch: false,
-        }
-    }
-
-    pub fn with_server(
-        mut self,
-        url: Option<&str>,
-        token: Option<&str>,
-        user_id: Option<&str>,
-    ) -> Self {
-        self.server = url.map(|url| LyricsServerAuth {
-            url: url.to_string(),
-            token: token.map(ToString::to_string),
-            user_id: user_id.map(ToString::to_string),
-        });
-        self
-    }
-
-    pub fn prefer_local(mut self, value: bool) -> Self {
-        self.prefer_local = value;
-        self
-    }
-
-    pub fn enable_musixmatch(mut self, value: bool) -> Self {
-        self.enable_musixmatch = value;
-        self
-    }
-
-    fn cache_key(&self) -> String {
-        lyrics_cache_key(
-            &self.artist,
-            &self.title,
-            &self.album,
-            self.duration,
-            &self.track_path,
-            self.enable_musixmatch,
-        )
-    }
-}
-
-const LYRICS_CACHE_CAPACITY: usize = 256;
 const MUSIXMATCH_ROOT_URL: &str = "https://apic-desktop.musixmatch.com/ws/1.1/";
 const PAXSENIX_ROOT_URL: &str = "https://lyrics.paxsenix.org";
 const ITUNES_SEARCH_ROOT_URL: &str = "https://itunes.apple.com/search";
@@ -131,8 +35,6 @@ const SERVER_LYRICS_TIMEOUT: Duration = Duration::from_secs(5);
 const LYRICS_INFLIGHT_POLL_INTERVAL: Duration = Duration::from_millis(50);
 const LYRICS_INFLIGHT_WAIT_TIMEOUT: Duration = Duration::from_secs(20);
 
-static LYRICS_CACHE: OnceLock<Mutex<LyricsCache>> = OnceLock::new();
-static LYRICS_INFLIGHT: OnceLock<Mutex<HashSet<String>>> = OnceLock::new();
 static MUSIXMATCH_TOKEN: OnceLock<Mutex<Option<MusixmatchToken>>> = OnceLock::new();
 
 macro_rules! lyrics_debug {
@@ -143,135 +45,9 @@ macro_rules! lyrics_debug {
     };
 }
 
-struct LyricsCache {
-    entries: HashMap<String, Option<Lyrics>>,
-    order: VecDeque<String>,
-    capacity: usize,
-}
-
-struct LyricsInflightGuard {
-    key: String,
-}
-
 struct MusixmatchToken {
     value: String,
     expires_at_ms: u128,
-}
-
-impl Drop for LyricsInflightGuard {
-    fn drop(&mut self) {
-        if let Ok(mut inflight) = lyrics_inflight().lock() {
-            inflight.remove(&self.key);
-        }
-    }
-}
-
-impl LyricsCache {
-    fn new(capacity: usize) -> Self {
-        Self {
-            entries: HashMap::new(),
-            order: VecDeque::new(),
-            capacity,
-        }
-    }
-
-    fn get_cloned(&mut self, key: &str) -> Option<Option<Lyrics>> {
-        let value = self.entries.get(key).cloned()?;
-        self.touch(key);
-        Some(value)
-    }
-
-    fn put(&mut self, key: String, value: Option<Lyrics>) {
-        if self.entries.contains_key(&key) {
-            self.entries.insert(key.clone(), value);
-            self.touch(&key);
-            return;
-        }
-
-        if self.entries.len() >= self.capacity {
-            while let Some(oldest) = self.order.pop_front() {
-                if self.entries.remove(&oldest).is_some() {
-                    break;
-                }
-            }
-        }
-
-        self.order.push_back(key.clone());
-        self.entries.insert(key, value);
-    }
-
-    fn touch(&mut self, key: &str) {
-        if let Some(pos) = self.order.iter().position(|existing| existing == key) {
-            self.order.remove(pos);
-        }
-        self.order.push_back(key.to_string());
-    }
-}
-
-// --- Persistent layer (metadata cache, kind="lyrics") ---------------------
-// The in-memory LRU stays the hot path; the DB makes lyrics survive restarts.
-// Negative results carry a TTL so a song that gains lyrics on a provider
-// later isn't permanently "no lyrics".
-
-const LYRICS_META_KIND: &str = "lyrics";
-const NEGATIVE_TTL_SECS: u64 = 7 * 24 * 60 * 60;
-
-fn now_unix() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0)
-}
-
-fn lyrics_to_payload(value: &Option<Lyrics>) -> String {
-    let v = match value {
-        // "synced2": full word-synced LyricLine shape. The pre-word-sync
-        // "synced" payloads parse as a miss below and simply refetch once.
-        Some(Lyrics::Synced(lines)) => serde_json::json!({
-            "kind": "synced2",
-            "lines": lines,
-        }),
-        Some(Lyrics::Plain(text)) => serde_json::json!({ "kind": "plain", "text": text }),
-        None => serde_json::json!({ "kind": "none", "ts": now_unix() }),
-    };
-    v.to_string()
-}
-
-fn lyrics_from_payload(payload: &str) -> Option<Option<Lyrics>> {
-    let v: serde_json::Value = serde_json::from_str(payload).ok()?;
-    match v.get("kind").and_then(|k| k.as_str())? {
-        "synced2" => {
-            let lines: Vec<LyricLine> = serde_json::from_value(v.get("lines")?.clone()).ok()?;
-            Some(Some(Lyrics::Synced(lines)))
-        }
-        "plain" => Some(Some(Lyrics::Plain(v.get("text")?.as_str()?.to_string()))),
-        "none" => {
-            let ts = v.get("ts").and_then(|t| t.as_u64()).unwrap_or(0);
-            if now_unix().saturating_sub(ts) < NEGATIVE_TTL_SECS {
-                Some(None) // fresh negative hit — skip the provider chain
-            } else {
-                None // stale negative — re-fetch
-            }
-        }
-        _ => None,
-    }
-}
-
-async fn load_persisted_lyrics(cache_key: &str) -> Option<Option<Lyrics>> {
-    let handle = crate::db_cache::get()?;
-    let payload = handle.meta_get(cache_key, LYRICS_META_KIND).await.ok()??;
-    lyrics_from_payload(&payload)
-}
-
-/// Store a fetch result in both layers: the in-memory LRU and the DB.
-async fn store_lyrics(cache_key: &str, value: &Option<Lyrics>) {
-    if let Ok(mut cache) = lyrics_cache().lock() {
-        cache.put(cache_key.to_string(), value.clone());
-    }
-    if let Some(handle) = crate::db_cache::get() {
-        let payload = lyrics_to_payload(value);
-        let _ = handle.meta_put(cache_key, LYRICS_META_KIND, &payload).await;
-    }
 }
 
 // --- Jellyfin lyrics types ---
@@ -1073,27 +849,6 @@ fn lyrics_terminal_debug_enabled() -> bool {
     #[cfg(not(target_arch = "wasm32"))]
     {
         std::env::var_os("KOPUZ_LYRICS_DEBUG").is_some()
-    }
-}
-
-fn lyrics_cache() -> &'static Mutex<LyricsCache> {
-    LYRICS_CACHE.get_or_init(|| Mutex::new(LyricsCache::new(LYRICS_CACHE_CAPACITY)))
-}
-
-fn lyrics_inflight() -> &'static Mutex<HashSet<String>> {
-    LYRICS_INFLIGHT.get_or_init(|| Mutex::new(HashSet::new()))
-}
-
-fn try_begin_lyrics_fetch(key: &str) -> bool {
-    let Ok(mut inflight) = lyrics_inflight().lock() else {
-        return true;
-    };
-
-    if inflight.contains(key) {
-        false
-    } else {
-        inflight.insert(key.to_string());
-        true
     }
 }
